minor change for testing

+++ src/database/migrations/thirdParties/2023_01_05_015929_create_logger_access_table.php
@@ -14,6 +14,7 @@ return new class extends Migration
      */
     public function up()
     {
+        if (!env('ACCESS_LOGGER_ENABLED')) return;
         $connection = env('TELESCOPE_DB_CONNECTION', 'mysql');
         Schema::connection($connection)->create('logger_access', function (Blueprint $table) {
             $table->id();
@@ -43,6 +44,7 @@ return new class extends Migration
      */
     public function down()
     {
+        if (!env('ACCESS_LOGGER_ENABLED')) return;
         $connection = env('TELESCOPE_DB_CONNECTION', 'mysql');
         Schema::connection($connection)->dropIfExists('logger_access');
     }
